Correct casts, and other small improvements
We now use static_cast when casting void* back to T*, instead of
reinterpret_cast; static_cast is the correct cast for this. Furthermore,
in the constructor, we now first static_cast the given T* to a void*
before using reinterpret_cast to cast that void* to our uintptr_t.
This is because uintptr_t is only guaranteed to be able to represent any
void*, and not necessarily any T* (and there is no requirement that
void* and T* have the same bit-level representation)! See the added
comments for a detailed explanation. We also introduce the
ContainsType<T, Ts...> concept, and add it as a constraint to all member
functions of TaggedPointer where it is applicable. Finally, we make
the get_tag_of_type() and num_types() methods static, and shorten the
implementation of get_tag_of_type().